chore: remove obsolete sarif renderer
